Fix exit calls in mkchroot; add fs_bind call
For some reason, 'exit 2' was being called upon display of an error message.
This inhibits cleanup actions as they had to be called before the error
message was shown.

Put an fs_bind call in when using a prebuilt image so we aren't constantly
downloading the same packages for every build.

Signed-off-by: Dan McGee <dan@archlinux.org>